New arch's single bus
Collapse all event-specific output into a single master bus. This
effectively make OpenDA Panel not a embedded patch anymore, but an
independent one, which plugs in with the user's patches. A collateral
API change is that initializing the engine now requires a list of paths
of where to find the patches, nad it must include ODA_PATCHES_PATH.